+++ test/unit/aws_encryption_sdk/key_providers/raw/test_raw_master_key_provider.py
@@ -18,7 +18,7 @@ from mock import MagicMock, patch, sentinel
 from aws_encryption_sdk.key_providers.base import MasterKeyProvider, MasterKeyProviderConfig
 from aws_encryption_sdk.key_providers.raw import RawMasterKeyProvider
 
-from ....test_values import VALUES
+from ....vectors import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
