+++ test/unit/aws_encryption_sdk/keyrings/raw/test_raw_rsa.py
@@ -23,7 +23,7 @@ from aws_encryption_sdk.internal.crypto.wrapping_keys import WrappingKey
 from aws_encryption_sdk.keyrings.base import Keyring
 from aws_encryption_sdk.keyrings.raw import RawRSAKeyring
 
-from ....test_values import VALUES
+from ....vectors import VALUES
 from ....unit_test_utils import (
     _BACKEND,
     _DATA_KEY,
