+++ test/unit/aws_encryption_sdk/streaming_client/test_stream_decryptor.py
@@ -22,7 +22,7 @@ from aws_encryption_sdk.key_providers.base import MasterKeyProvider
 from aws_encryption_sdk.materials_managers.base import CryptoMaterialsManager
 from aws_encryption_sdk.streaming_client import StreamDecryptor
 
-from ...test_values import VALUES
+from ...vectors import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
